requirements.txt
flask
line-bot-sdk==3.4.0
requests
geopy
python-dotenv
